Maven Project=>Implementatio of Spring MVC
In this Application Dispatcher servlet delegate the request to handler mapping and find Handler find a method in Controller for the mapping by @RequetMapping.
Then, this method return the view to Dispatcher servlet and then send to the ViewResolver for the view Then finally display the view either is JSP or HTML (including Css,Ajax,Javascript and jQuery)
Controller may also return the Model modal whenever requierd.